requirements.txt
numpy
opencv-python
pandas
scikit-learn
onnxruntime
pyside6
gspread
google-auth
pillow
imutils
pyserial
